Move `read_lines` to `resources.py`
